requirements.txt
celery
eventlet
Flask
Flask-Bcrypt
Flask-Caching
Flask-Compress
Flask-Cors
Flask-JWT-Extended
Flask-Login
Flask-Mail
flask-marshmallow
Flask-SocketIO
Flask-SQLAlchemy
flask-talisman
Flask-WTF
gunicorn
itsdangerous
Jinja2
MarkupSafe
marshmallow-sqlalchemy
Pillow
psycopg2-binary
py2neo
py4j
redis
shortuuid
SQLAlchemy
validators
watchdog
websocket-client
Werkzeug
WTForms
email-validator
bidict
kombu 


